requirements.txt
pillow==8.3.2
numpy==1.21.2
streamlit==1.32.0